+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Change
+
+- Output Json's format changed. Now `"fruits"` & `"details"` are named as `"test_targets"` & `"test_cases"`, `"is_expected"` is replaced by `"expect"`, etc.
+
 ## [0.8.1] - 2021-06-17
 
 ### Change